fix: fix kodiak quoter interface

+++ contracts/integrations/kodiak/IKodiakQuoter.sol
@@ -5,7 +5,7 @@ struct QuoteExactInputSingleParams {
     address tokenIn;
     address tokenOut;
     uint256 amountIn;
-    uint256 fee;
+    uint24 fee;
     uint160 sqrtPriceLimitX96;
 }
